more compatibility fixes for python2/3
