+++ tokio/common.py
@@ -1,6 +1,11 @@
 """Common convenience routines used throughout pytokio
 """
 
+import time
+import json
+import datetime
+import numpy
+
 def isstr(obj):
     """Determine if an object is a string or string-derivative
 
@@ -18,3 +23,13 @@ def isstr(obj):
     except NameError:
         # Python 3
         return isinstance(obj, str)
+
+class JSONEncoder(json.JSONEncoder):
+    """Convert common pytokio data types into serializable formats
+    """
+    def default(self, obj): # pylint: disable=E0202
+        if isinstance(obj, datetime.datetime):
+            return int(time.mktime(obj.timetuple()))
+        elif isinstance(obj, numpy.int64):
+            return int(obj)
+        return json.JSONEncoder.default(self, obj)
